Add a new type length_type in types.h (typedef for u64bit)
In nearly every place where a u32bit was used, and was not specifically
required to be a u32bit (for instance, when passing the length of a buffer),
instead use length_type.

It's not clear at this point if this should actually be a distinct typedef,
or if we should just be using size_t. If size_t does turn out to make
more sense, it is the work of a simple Perl one-liner to do the conversion;
the hard part was examining each declaration of a u32bit and seeing how
it was used.